# Conflicts:
#	nekoyume/Assets/AddressableAssets/UI/Module/ItemView/ItemView.prefab
#	nekoyume/Assets/_Scenes/Game.unity
#	nekoyume/Assets/_Scripts/BlockChain/ActionManager.cs
#	nekoyume/Assets/_Scripts/BlockChain/ActionRenderHandler.cs

+++ nekoyume/Assets/_Scripts/UI/BlockTimer.cs
@@ -49,7 +49,7 @@ namespace Nekoyume.UI
 
         private void OnSliderChange(float value)
         {
-            var time = Util.GetBlockToTime((int) value);
+            var time = Util.GetBlockToTime((long) value);
             remainTime.text = string.Format(L10nManager.Localize("UI_BLOCK_TIMER"), value, time);
         }
